\name{etlKoreanNSC-package}
\alias{etlKoreanNSC-package}
\alias{etlKoreanNSC}
\docType{package}
\title{
\packageTitle{etlKoreanNSC}
}
\description{
\packageDescription{etlKoreanNSC}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{etlKoreanNSC}
\packageIndices{etlKoreanNSC}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{etlKoreanNSC}

Maintainer: \packageMaintainer{etlKoreanNSC}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
