% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executeETL.R
\name{executeNHISETL}
\alias{executeNHISETL}
\title{executeETL Function}
\usage{
executeNHISETL(NHISNSC_rawdata, NHISNSC_database, Mapping_database,
  NHIS_JK, NHIS_20T, NHIS_30T, NHIS_40T, NHIS_60T, NHIS_GJ, NHIS_YK,
  connection, outputFolder, CDM_ddl = FALSE, master_table = FALSE,
  location = FALSE, care_site = FALSE, person = FALSE,
  death = FALSE, observation_period = FALSE,
  visit_occurrence = FALSE, condition_occurrence = FALSE,
  observation = FALSE, drug_exposure = TRUE,
  procedure_occurrence = TRUE, device_exposure = TRUE,
  measurement = TRUE, payer_plan_period = TRUE, cost = TRUE,
  generateEra = TRUE, dose_era = TRUE, cdm_source = TRUE,
  indexing = TRUE, constraints = TRUE, data_cleansing = TRUE)
}
\arguments{
\item{NHISNSC_rawdata}{raw schema}

\item{NHISNSC_database}{load schema}

\item{Mapping_database}{mapping schema}

\item{NHIS_JK}{JK table}

\item{NHIS_20T}{20T table}

\item{NHIS_30T}{30T table}

\item{NHIS_40T}{40T table}

\item{NHIS_60T}{60T table}

\item{NHIS_GJ}{GJ table}

\item{NHIS_YK}{YK table}

\item{connection}{db connection information}

\item{outputFolder}{error load path}

\item{CDM_ddl}{boolean}

\item{master_table}{boolean}

\item{location}{boolean}

\item{care_site}{boolean}

\item{person}{boolean}

\item{death}{boolean}

\item{observation_period}{boolean}

\item{visit_occurrence}{boolean}

\item{condition_occurrence}{boolean}

\item{observation}{boolean}

\item{drug_exposure}{boolean}

\item{procedure_occurrence}{boolean}

\item{device_exposure}{boolean}

\item{measurement}{boolean}

\item{payer_plan_period}{boolean}

\item{cost}{boolean}

\item{generateEra}{boolean}

\item{dose_era}{boolean}

\item{cdm_source}{boolean}

\item{indexing}{boolean}

\item{constraints}{boolean}

\item{data_cleansing}{boolean}
}
\description{
This function allows to execute ETL process.
}
